utils/ai_engine.py
import requests
from bs4 import BeautifulSoup

def scrape_url_content(url):
    try:
        response = requests.get(url)
        response.raise_for_status()  # Raise an exception for HTTP errors
        soup = BeautifulSoup(response.text, 'html.parser')
        # Extract readable content (e.g., paragraphs, article body)
        paragraphs = soup.find_all('p')
        content = '\n'.join([p.get_text() for p in paragraphs])
        return content
    except requests.exceptions.RequestException as e:
        print(f"Error scraping URL {url}: {e}")
        return None

def generate_content_with_ai(text_input, ai_model):
    # Placeholder for actual AI model integration
    # In a real scenario, you would call the respective AI API (ChatGPT, Gemini, etc.)
    print(f"Generating content with AI model {ai_model} for input: {text_input[:100]}...")
    return f"AI generated content based on {text_input}"

